+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - PR #2211: MNMG KNN Classifier & Regressor
 - PR #2461: Add KNN Sparse Output Functionality
 - PR #2615: Incremental PCA
+- PR #2607: Add support for probability estimates in SVC
 
 ## Improvements
 - PR #2336: Eliminate `rmm.device_array` usage
@@ -112,6 +113,9 @@
 - PR #2586: Fix SVC decision function data type
 - PR #2573: Considering managed memory as device type on checking for KMeans
 - PR #2574: Fixing include path in `tsvd_mg.pyx`
+- PR #2593: Fix inconsistency in train_test_split
+- PR #2609: Fix small doxygen issues
+- PR #2610: Remove cuDF tolist call
 
 # cuML 0.14.0 (03 Jun 2020)